refactor(Navigation): moving ParamsList type into Navigation module, updating type imports where applies, updated version number and release notes
