+++ src/components/LabelledHint/LabelledHint.styles.ts
@@ -5,7 +5,9 @@ const styles = {
     paddingVertical: 8,
     flexDirection: 'row',
   } as ViewStyle,
-  icon: { marginRight: 4 } as ViewStyle,
+  icon: {
+    marginRight: 4,
+  } as ViewStyle,
 }
 
 export default StyleSheet.create(styles)
