+++ src/modules/Landing/Landing.types.ts
@@ -1,5 +1,3 @@
-import { HomeNavigatorParamsList } from '~/modules/HomeNavigation/HomeNavigation.types'
-
 export enum LandingNavigationRoutes {
   Authentication = 'Authentication',
 }
@@ -7,5 +5,3 @@ export enum LandingNavigationRoutes {
 export type LandingParamsList = {
   [LandingNavigationRoutes.Authentication]: undefined
 }
-
-export type ParamsList = LandingParamsList | HomeNavigatorParamsList
