+++ src/modules/Navigation/Navigation.types.ts
@@ -1,0 +1,4 @@
+import { HomeNavigatorParamsList } from '../HomeNavigation/HomeNavigation.types'
+import { LandingParamsList } from '../Landing/Landing.types'
+
+export type ParamsList = LandingParamsList | HomeNavigatorParamsList
