+++ src/modules/Navigation/Navigation.utils.ts
@@ -1,6 +1,6 @@
 import { createNavigationContainerRef } from '@react-navigation/native'
 
-import { ParamsList } from '~/modules/Landing/Landing.types'
+import { ParamsList } from './Navigation.types'
 
 export const navigationRef = createNavigationContainerRef<ParamsList>()
