+++ src/modules/Navigation/index.ts
@@ -1,4 +1,5 @@
+import * as types from './Navigation.types'
 import * as utils from './Navigation.utils'
 
 export { default } from './Navigation.container'
-export { utils }
+export { utils, types }
