+++ src/modules/UserChallenges/UserChallenges.middleware.ts
@@ -8,7 +8,7 @@ import { constants as ApiUsersConstants, types as ApiUsersTypes } from '~/api/us
 import * as Strings from '~/constants/strings.constants'
 import { extractErrorResponseMessage } from '~/modules/Error/error.utils'
 import { HomeNavigationRoutes } from '~/modules/HomeNavigation/HomeNavigation.types'
-import * as Navigation from '~/modules/Navigation/Navigation.utils'
+import { types as NavigationTypes, utils as NavigationUtils } from '~/modules/Navigation'
 import {
   actions as UserActions,
   constants as UserConstants,
@@ -21,7 +21,6 @@ import * as ReduxUtils from '~/redux/redux.utils'
 import * as Types from '~/types/general.types'
 import * as ErrorUtils from '~/utils/error'
 
-import { ParamsList } from '../Landing/Landing.types'
 import {
   createUserChallenge,
   createUserChallengeCertificate,
@@ -83,7 +82,7 @@ export const createUserChallengeSuccessFlow =
     navigate,
   }: {
     notification: typeof ErrorUtils.showSimpleMessage
-    navigate: typeof Navigation.navigate
+    navigate: typeof NavigationUtils.navigate
   }): Middleware =>
   ({ dispatch }) =>
   next =>
@@ -94,7 +93,7 @@ export const createUserChallengeSuccessFlow =
       const userChallenge = ReduxUtils.extractDataFromResponseAction(action)
       const normalisedUserChallenge = ReduxUtils.normalise(of(userChallenge))
       notification('success', i18n.t(Strings.NEW_CHALLENGE_SUCCESSFULLY_CREATED))
-      navigate(HomeNavigationRoutes.Home as keyof ParamsList)
+      navigate(HomeNavigationRoutes.Home as keyof NavigationTypes.ParamsList)
       dispatch(updateUserChallenges(normalisedUserChallenge))
       dispatch(createUserChallengeCertificate(userChallenge.id))
     }
